requirements.txt
fastapi
uvicorn
openai
aquiles-rag
Artemisa
libsql
python-dotenv
platformdirs